web/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0
python-dotenv>=1.0.0
ollama>=0.4.0
python-multipart>=0.0.9
